Updated class structure with abstract supabaseStorage class
